index.php
This is your main website file. Make sure it has .php at the end. This code is meant for Bungee.
You would replace us.mineplex.com, eu.mineplex.com, and us.the-nebula.net with your servers in your Bungee network.

This DOES NOT work with ports. PHP code with ports is releasing soon.
Wherever you want to put the live number of players, put <?php echo $players; ?> . Whether that be in a navbar, in the middle of your webpage, it doesn't matter. The result is the same.